requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.8.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
werkzeug==3.0.1
markupsafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
jinja2==3.1.2
cryptography==41.0.7
bcrypt==4.1.1
pytz==2023.3
python-dateutil==2.8.2
requests==2.31.0
